LinkedIn : Given two strings, determine if they are isomorphic.
Two words are called isomorphic if the letters in one word can be
remapped to get the second word. Remapping a letter means replacing all
occurrences of it with another letter while the ordering of the letters
remains unchanged. No two letters may map to the same letter, but a
letter may map to itself.